add job for nokogiri
