+++ config/application.rb
@@ -27,5 +27,6 @@ module RailsDevToolsLinks
         resource '*', :headers => :any, :methods => [:get, :post, :options]
       end
     end
+    config.active_job.queue_adapter = :sidekiq
   end
 end
