+++ config/initializers/redis.rb
@@ -1,0 +1,14 @@
+$redis = Redis.new
+
+url = ENV["REDISCLOUD_URL"]
+
+if url
+  Sidekiq.configure_server do |config|
+    config.redis = { url: url }
+  end
+
+  Sidekiq.configure_client do |config|
+    config.redis = { url: url }
+  end
+  $redis = Redis.new(:url => url)
+end
