+++ config/sidekiq.yml
@@ -1,0 +1,8 @@
+:concurrency: 3
+:timeout: 60
+:verbose: true
+:queues:
+# Queue priority:
+# https://github.com/mperham/sidekiq/wiki/Advanced-Options
+  - default
+  - mailers
