+++ test/jobs/nokogiri_image_scrapping_job_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class NokogiriImageScrappingJobTest < ActiveJob::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
